text_to_image_app/requirements.txt
streamlit
fastapi
uvicorn
torch
transformers
diffusers
accelerate
python-dotenv
pillow
python-multipart
requests